(Fix 2) Fixed issues with sample projects (CDA.R5Samples & CDA.Sample) where they would both fail if the user provided an OutputFolder as an argument to the console project, in order to have the output content to go to a specified folder, rather than into the applications / bin folder.
When this was attempted the application would always fail on loading the Logo.png file, further more in the CDA.Sample project it would also fail when calling the 'LevelsGenerator.Generate2(inputPath, outputPath)' method as the users custom OutputFolder was not considered in this method call.
To resolve all this I have rewritten the method 'LogoSetupAndValidation' found in the CDAGenerator class and also updated every sample to explicitly set the 'LogoPath' to the users custom OutputFolder, for example 'serviceReferral.LogoPath = OutputFolderPath', the case where no custom OutputFolder is provided is also still covered.
I have also created a support method in the CDA.Sample project's root Program class called LevelsGeneratorPathCorrections(string input, string output, string OutPutFolderPath) which manages updating these paths to also consider the users custom OutputFolder.
I have then tested that all works for both sample project both when the users custom OutputFolder is provided and when it is not, in both cases the project now work correctly.